feat(mobile): merge search screen into discover screnn; optimize search input
