+++ apps/mobile/src/modules/discover/search.tsx
@@ -1,9 +1,7 @@
 import { useAtom, useAtomValue, useSetAtom } from "jotai"
-import type { FC } from "react"
 import { useContext, useEffect, useRef, useState } from "react"
 import { useTranslation } from "react-i18next"
-import type { Animated as RnAnimated, LayoutChangeEvent } from "react-native"
-import { Pressable, StyleSheet, Text, TextInput, TouchableOpacity, View } from "react-native"
+import { StyleSheet, Text, TextInput, View } from "react-native"
 import Animated, {
   Easing,
   useAnimatedStyle,
@@ -12,44 +10,16 @@ import Animated, {
 } from "react-native-reanimated"
 import { useSafeAreaFrame, useSafeAreaInsets } from "react-native-safe-area-context"
 
+import { ReAnimatedTouchableOpacity } from "@/src/components/common/AnimatedComponents"
 import { BlurEffect } from "@/src/components/common/BlurEffect"
 import { getDefaultHeaderHeight } from "@/src/components/layouts/utils"
 import { Search2CuteReIcon } from "@/src/icons/search_2_cute_re"
 import { useScreenIsInSheetModal } from "@/src/lib/navigation/hooks"
-import { Navigation } from "@/src/lib/navigation/Navigation"
 import { ScreenItemContext } from "@/src/lib/navigation/ScreenItemContext"
-import SearchScreen from "@/src/screens/(headless)/search"
 import { accentColor, useColor } from "@/src/theme/colors"
 
 import { useSearchPageContext } from "./ctx"
 import { DiscoverContext } from "./DiscoverContext"
-import { SearchTabBar } from "./SearchTabBar"
-
-export const SearchHeader: FC<{
-  animatedX: RnAnimated.Value
-  onLayout: (e: LayoutChangeEvent) => void
-}> = ({ animatedX, onLayout }) => {
-  const frame = useSafeAreaFrame()
-  const insets = useSafeAreaInsets()
-
-  const sheetModal = useScreenIsInSheetModal()
-  const headerHeight = getDefaultHeaderHeight(frame, sheetModal, insets.top)
-
-  return (
-    <View
-      style={{ minHeight: headerHeight, paddingTop: insets.top }}
-      className="relative"
-      onLayout={onLayout}
-    >
-      <BlurEffect />
-
-      <View style={styles.header}>
-        <ComposeSearchBar />
-      </View>
-      <SearchTabBar animatedX={animatedX} />
-    </View>
-  )
-}
 
 const DynamicBlurEffect = () => {
   const { reAnimatedScrollY } = useContext(ScreenItemContext)
@@ -83,58 +53,9 @@ export const DiscoverHeader = () => {
       <DynamicBlurEffect />
 
       <View style={styles.header}>
-        <PlaceholerSearchBar />
+        <SearchInput />
       </View>
     </View>
-  )
-}
-
-const PlaceholerSearchBar = () => {
-  const labelColor = useColor("secondaryLabel")
-  const { t } = useTranslation("common")
-  return (
-    <Pressable
-      style={styles.searchbar}
-      className="bg-tertiary-system-fill"
-      onPress={() => {
-        Navigation.rootNavigation.pushControllerView(SearchScreen)
-      }}
-    >
-      <View
-        className="absolute inset-0 flex flex-row items-center justify-center"
-        pointerEvents="none"
-      >
-        <Search2CuteReIcon color={labelColor} height={18} width={18} />
-        <Text className="text-secondary-label ml-1" style={styles.searchPlaceholderText}>
-          {t("words.search")}
-        </Text>
-      </View>
-    </Pressable>
-  )
-}
-
-const ComposeSearchBar = () => {
-  const { searchFocusedAtom, searchValueAtom } = useSearchPageContext()
-  const setIsFocused = useSetAtom(searchFocusedAtom)
-  const setSearchValue = useSetAtom(searchValueAtom)
-  return (
-    <>
-      <SearchInput />
-
-      <TouchableOpacity
-        hitSlop={10}
-        onPress={() => {
-          setIsFocused(false)
-          setSearchValue("")
-
-          if (Navigation.rootNavigation.canGoBack()) {
-            Navigation.rootNavigation.back()
-          }
-        }}
-      >
-        <Text className="text-accent ml-3 text-lg font-medium">Cancel</Text>
-      </TouchableOpacity>
-    </>
   )
 }
 
@@ -149,7 +70,9 @@ const SearchInput = () => {
 
   const skeletonOpacity = useSharedValue(0)
   const skeletonTranslateX = useSharedValue(0)
-  const placeholderOpacity = useSharedValue(1)
+  const placeholderOpacity = useSharedValue(0)
+  const marginRight = useSharedValue(0)
+  const cancelButtonTranslateX = useSharedValue(20)
 
   const [tempSearchValue, setTempSearchValue] = useState(searchValue)
 
@@ -157,21 +80,32 @@ const SearchInput = () => {
 
   useEffect(() => {
     if (focusOrHasValue) {
-      skeletonOpacity.value = withTiming(0, { duration: 100, easing: Easing.ease })
+      skeletonOpacity.value = withTiming(0, { duration: 100 })
       skeletonTranslateX.value = withTiming(-150, {
         duration: 100,
         easing: Easing.inOut(Easing.ease),
       })
-      placeholderOpacity.value = withTiming(1, { duration: 200, easing: Easing.ease })
+      placeholderOpacity.value = withTiming(1, { duration: 200 })
+      marginRight.value = withTiming(64, { duration: 200 })
+      cancelButtonTranslateX.value = withTiming(0, { duration: 200 })
     } else {
-      skeletonOpacity.value = withTiming(1, { duration: 100, easing: Easing.ease })
+      skeletonOpacity.value = withTiming(1, { duration: 100 })
       skeletonTranslateX.value = withTiming(0, {
         duration: 100,
         easing: Easing.inOut(Easing.ease),
       })
-      placeholderOpacity.value = withTiming(0, { duration: 200, easing: Easing.ease })
+      placeholderOpacity.value = withTiming(0, { duration: 200 })
+      marginRight.value = withTiming(0, { duration: 200 })
+      cancelButtonTranslateX.value = withTiming(20, { duration: 200 })
     }
-  }, [focusOrHasValue, placeholderOpacity, skeletonOpacity, skeletonTranslateX])
+  }, [
+    focusOrHasValue,
+    placeholderOpacity,
+    skeletonOpacity,
+    skeletonTranslateX,
+    marginRight,
+    cancelButtonTranslateX,
+  ])
 
   const skeletonAnimatedStyle = useAnimatedStyle(() => ({
     opacity: skeletonOpacity.value,
@@ -198,6 +132,14 @@ const SearchInput = () => {
     justifyContent: "center",
   }))
 
+  const containerAnimatedStyle = useAnimatedStyle(() => ({
+    marginRight: marginRight.value,
+  }))
+
+  const cancelButtonAnimatedStyle = useAnimatedStyle(() => ({
+    transform: [{ translateX: cancelButtonTranslateX.value }],
+  }))
+
   useEffect(() => {
     if (!isFocused) {
       inputRef.current?.blur()
@@ -207,44 +149,58 @@ const SearchInput = () => {
   }, [isFocused])
 
   return (
-    <View style={styles.searchbar} className="bg-tertiary-system-fill">
-      {focusOrHasValue && (
-        <Animated.View style={placeholderAnimatedStyle}>
+    <Animated.View className="flex-row items-center justify-center" style={containerAnimatedStyle}>
+      <View style={styles.searchbar} className="bg-tertiary-system-fill">
+        {focusOrHasValue && (
+          <Animated.View style={placeholderAnimatedStyle}>
+            <Search2CuteReIcon color={placeholderTextColor} height={18} width={18} />
+            {!searchValue && !tempSearchValue && (
+              <Text className="text-secondary-label ml-2" style={styles.searchPlaceholderText}>
+                {t("words.search")}
+              </Text>
+            )}
+          </Animated.View>
+        )}
+        <TextInput
+          enterKeyHint="search"
+          autoFocus={isFocused}
+          ref={inputRef}
+          onSubmitEditing={() => {
+            setSearchValue(tempSearchValue)
+          }}
+          defaultValue={searchValue}
+          cursorColor={accentColor}
+          selectionColor={accentColor}
+          style={styles.searchInput}
+          className="text-text"
+          onFocus={() => setIsFocused(true)}
+          onBlur={() => !searchValue && !tempSearchValue && setIsFocused(false)}
+          onChangeText={(text) => {
+            setTempSearchValue(text)
+          }}
+        />
+
+        <Animated.View style={skeletonAnimatedStyle} pointerEvents="none">
           <Search2CuteReIcon color={placeholderTextColor} height={18} width={18} />
-          {!searchValue && !tempSearchValue && (
-            <Text className="text-secondary-label ml-2" style={styles.searchPlaceholderText}>
-              {t("words.search")}
-            </Text>
-          )}
+          <Text className="text-secondary-label ml-1" style={styles.searchPlaceholderText}>
+            {t("words.search")}
+          </Text>
         </Animated.View>
-      )}
-      <TextInput
-        enterKeyHint="search"
-        autoFocus={isFocused}
-        ref={inputRef}
-        onSubmitEditing={() => {
-          setSearchValue(tempSearchValue)
+      </View>
+
+      <ReAnimatedTouchableOpacity
+        hitSlop={10}
+        onPress={() => {
+          setIsFocused(false)
+          setSearchValue("")
           setTempSearchValue("")
         }}
-        defaultValue={searchValue}
-        cursorColor={accentColor}
-        selectionColor={accentColor}
-        style={styles.searchInput}
-        className="text-text"
-        onFocus={() => setIsFocused(true)}
-        onBlur={() => setIsFocused(false)}
-        onChangeText={(text) => {
-          setTempSearchValue(text)
-        }}
-      />
-
-      <Animated.View style={skeletonAnimatedStyle} pointerEvents="none">
-        <Search2CuteReIcon color={placeholderTextColor} height={18} width={18} />
-        <Text className="text-secondary-label ml-1" style={styles.searchPlaceholderText}>
-          {t("words.search")}
-        </Text>
-      </Animated.View>
-    </View>
+        className="absolute -right-20 w-20 pl-4"
+        style={cancelButtonAnimatedStyle}
+      >
+        <Text className="text-accent text-lg font-medium">Cancel</Text>
+      </ReAnimatedTouchableOpacity>
+    </Animated.View>
   )
 }
 const styles = StyleSheet.create({
@@ -255,6 +211,7 @@ const styles = StyleSheet.create({
     marginBottom: 14,
     marginHorizontal: 16,
     position: "relative",
+    marginTop: 4,
   },
 
   searchbar: {
